# docs/openapi.yaml
openapi: 3.1.0
info:
  title: 山东大学黑客松 API
  version: 0.0.1
  description: |
    根据仓库中的 Spring MVC Controller、DTO/VO、校验注解、鉴权拦截器与业务状态码整理。
    除 SDU Pass 回调外，业务接口统一返回 `{code, data, msg}`；成功时 `code` 为 200。
    受保护接口使用访问令牌，服务端同时兼容 `Authorization: Bearer <token>` 与直接传入令牌，本文档按标准 Bearer 方式描述。
    `x-required-roles` 和 `x-required-window` 是对仓库 `@Require` 约束的补充说明。
servers:
  - url: http://localhost:8080
    description: 本地开发环境
...
paths:
...
  /team/{id}:
    parameters:
      - {name: id, in: path, required: true, schema: {type: integer, format: int32, minimum: 1}, description: POST 时为赛道 ID；GET、PUT、DELETE 时为队伍 ID}
    post:
      tags: [队伍]
      operationId: createTeam
      summary: 在赛道下创建队伍
      security: [{bearerAuth: []}]
      x-required-roles: [STUDENT]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/CreateTeamDTO'}}}
      responses:
        '200': {description: 创建成功, content: {application/json: {schema: {$ref: '#/components/schemas/TeamIdResult'}}}}
        default: {$ref: '#/components/responses/ErrorResponse'}
    get:
      tags: [队伍]
      operationId: getTeam
      summary: 查询队伍详情
      security: [{bearerAuth: []}]
      x-required-roles: [TEAM_MEMBER, EVENT_ADMIN]
      responses:
        '200': {description: 获取成功, content: {application/json: {schema: {$ref: '#/components/schemas/TeamInfoResult'}}}}
        default: {$ref: '#/components/responses/ErrorResponse'}
    put:
      tags: [队伍]
      operationId: updateTeam
      summary: 修改队伍信息
      security: [{bearerAuth: []}]
      x-required-roles: [TEAM_LEADER, EVENT_ADMIN]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/UpdateTeamDTO'}}}
      responses:
        '200': {$ref: '#/components/responses/OkResponse'}
        default: {$ref: '#/components/responses/ErrorResponse'}
    delete:
      tags: [队伍]
      operationId: deleteTeam
      summary: 解散队伍
      security: [{bearerAuth: []}]
      x-required-roles: [TEAM_LEADER, EVENT_ADMIN]
      responses:
        '200': {$ref: '#/components/responses/OkResponse'}
        default: {$ref: '#/components/responses/ErrorResponse'}
  /team/{teamId}/join:
    parameters:
      - {$ref: '#/components/parameters/TeamId'}
    post:
      tags: [队伍]
      operationId: joinTeam
      summary: 加入队伍
      security: [{bearerAuth: []}]
      x-required-roles: [STUDENT]
      responses:
        '200': {$ref: '#/components/responses/OkResponse'}
        default: {$ref: '#/components/responses/ErrorResponse'}
  /team/{teamId}/invite:
    parameters:
      - {$ref: '#/components/parameters/TeamId'}
    post:
      tags: [队伍]
      operationId: inviteTeamMember
      summary: 向用户发送待确认的组队邀请
      security: [{bearerAuth: []}]
      x-required-roles: [TEAM_LEADER, EVENT_ADMIN]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/UserIdDTO'}}}
      responses:
        '200': {description: 邀请已发送, content: {application/json: {schema: {$ref: '#/components/schemas/IntegerResult'}}}}
        default: {$ref: '#/components/responses/ErrorResponse'}
  /team/{teamId}/leave:
    parameters:
      - {$ref: '#/components/parameters/TeamId'}
    delete:
      tags: [队伍]
      operationId: leaveTeam
      summary: 离开队伍
      security: [{bearerAuth: []}]
      x-required-roles: [TEAM_MEMBER]
      responses:
        '200': {$ref: '#/components/responses/OkResponse'}
        default: {$ref: '#/components/responses/ErrorResponse'}
  /team/{teamId}/kick:
    parameters:
      - {$ref: '#/components/parameters/TeamId'}
    delete:
      tags: [队伍]
      operationId: kickTeamMember
      summary: 移除队伍成员
      security: [{bearerAuth: []}]
      x-required-roles: [TEAM_LEADER, EVENT_ADMIN]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/UserIdDTO'}}}
      responses:
        '200': {$ref: '#/components/responses/OkResponse'}
        default: {$ref: '#/components/responses/ErrorResponse'}
...
  /team/join-code:
    post:
      tags: [队伍]
      operationId: joinTeamByInviteCode
      summary: 使用私密邀请码加入队伍
      security: [{bearerAuth: []}]
      x-required-roles: [STUDENT]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/InviteCodeDTO'}}}
      responses:
        '200': {$ref: '#/components/responses/OkResponse'}
        default: {$ref: '#/components/responses/ErrorResponse'}
  /team/{teamId}/invite-code:
    post:
      tags: [队伍]
      operationId: refreshTeamInviteCode
      summary: 刷新队伍私密邀请码，旧码立即失效
      security: [{bearerAuth: []}]
      x-required-roles: [TEAM_LEADER, EVENT_ADMIN]
      parameters: [{$ref: '#/components/parameters/TeamId'}]
      responses:
        '200': {description: 邀请码已刷新, content: {application/json: {schema: {$ref: '#/components/schemas/StringResult'}}}}
        default: {$ref: '#/components/responses/ErrorResponse'}
  /team/{teamId}/members:
    post:
      tags: [队伍]
      operationId: adminAddTeamMember
      summary: 赛事管理员人工添加队员
      security: [{bearerAuth: []}]
      x-required-roles: [EVENT_ADMIN]
      parameters: [{$ref: '#/components/parameters/TeamId'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/UserIdDTO'}}}
      responses:
        '200': {$ref: '#/components/responses/OkResponse'}
        default: {$ref: '#/components/responses/ErrorResponse'}
  /team/{teamId}/leader/transfer:
    post:
      tags: [队伍]
      operationId: transferTeamLeader
      summary: 当前队长将队长职责转让给本队成员
      security: [{bearerAuth: []}]
      x-required-roles: [TEAM_LEADER]
      parameters: [{$ref: '#/components/parameters/TeamId'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/UserIdDTO'}}}
      responses:
        '200': {$ref: '#/components/responses/OkResponse'}
        default: {$ref: '#/components/responses/ErrorResponse'}
  /team/{teamId}/leader:
    put:
      tags: [队伍]
      operationId: adminSetTeamLeader
      summary: 赛事管理员将本队成员设置为队长
      security: [{bearerAuth: []}]
      x-required-roles: [EVENT_ADMIN]
      parameters: [{$ref: '#/components/parameters/TeamId'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/UserIdDTO'}}}
      responses:
        '200': {$ref: '#/components/responses/OkResponse'}
        default: {$ref: '#/components/responses/ErrorResponse'}
  /team/invitations:
    get:
      tags: [队伍]
      operationId: listMyTeamInvitations
      summary: 查询本人收到的组队邀请
      security: [{bearerAuth: []}]
      x-required-roles: [STUDENT]
      responses:
        '200': {description: 获取成功, content: {application/json: {schema: {$ref: '#/components/schemas/TeamInvitationListResult'}}}}
        default: {$ref: '#/components/responses/ErrorResponse'}
  /team/{teamId}/invitations/{invitationId}/accept:
    post:
      tags: [队伍]
      operationId: acceptTeamInvitation
      summary: 接受组队邀请
      security: [{bearerAuth: []}]
      x-required-roles: [STUDENT]
      parameters:
        - {$ref: '#/components/parameters/TeamId'}
        - {name: invitationId, in: path, required: true, schema: {type: integer, format: int32}}
      responses:
        '200': {$ref: '#/components/responses/OkResponse'}
        default: {$ref: '#/components/responses/ErrorResponse'}
  /team/{teamId}/invitations/{invitationId}/reject:
    post:
      tags: [队伍]
      operationId: rejectTeamInvitation
      summary: 拒绝组队邀请
      security: [{bearerAuth: []}]
      x-required-roles: [STUDENT]
      parameters:
        - {$ref: '#/components/parameters/TeamId'}
        - {name: invitationId, in: path, required: true, schema: {type: integer, format: int32}}
      responses:
        '200': {$ref: '#/components/responses/OkResponse'}
        default: {$ref: '#/components/responses/ErrorResponse'}
  /team/{teamId}/recruitments:
    post:
      tags: [人才招募]
      operationId: createTeamRecruitment
      summary: 发布队伍人才缺口
      security: [{bearerAuth: []}]
      x-required-roles: [TEAM_LEADER, EVENT_ADMIN]
      parameters: [{$ref: '#/components/parameters/TeamId'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/CreateRecruitmentDTO'}}}
      responses:
        '200': {description: 发布成功, content: {application/json: {schema: {$ref: '#/components/schemas/IntegerResult'}}}}
        default: {$ref: '#/components/responses/ErrorResponse'}
  /team/{teamId}/recruitments/{recruitmentId}:
    parameters:
      - {$ref: '#/components/parameters/TeamId'}
      - {name: recruitmentId, in: path, required: true, schema: {type: integer, format: int32}}
    put:
      tags: [人才招募]
      operationId: updateTeamRecruitment
      summary: 修改或关闭人才缺口
      security: [{bearerAuth: []}]
      x-required-roles: [TEAM_LEADER, EVENT_ADMIN]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/UpdateRecruitmentDTO'}}}
      responses:
        '200': {$ref: '#/components/responses/OkResponse'}
        default: {$ref: '#/components/responses/ErrorResponse'}
    delete:
      tags: [人才招募]
      operationId: deleteTeamRecruitment
      summary: 删除人才缺口
      security: [{bearerAuth: []}]
      x-required-roles: [TEAM_LEADER, EVENT_ADMIN]
      responses:
        '200': {$ref: '#/components/responses/OkResponse'}
        default: {$ref: '#/components/responses/ErrorResponse'}
  /team/events/{eventId}/recruitments/list:
    post:
      tags: [人才招募]
      operationId: listRecruitments
      summary: 分页检索开放的队伍招募信息
      security: [{bearerAuth: []}]
      x-required-roles: [STUDENT]
      parameters:
        - {$ref: '#/components/parameters/EventId'}
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/Size'}
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/QueryRecruitmentDTO'}}}
      responses:
        '200': {description: 获取成功, content: {application/json: {schema: {$ref: '#/components/schemas/RecruitmentPageResult'}}}}
        default: {$ref: '#/components/responses/ErrorResponse'}
  /team/recruitments/{recruitmentId}/applications:
    post:
      tags: [人才招募]
      operationId: applyToTeam
      summary: 申请加入招募队伍
      security: [{bearerAuth: []}]
      x-required-roles: [STUDENT]
      parameters:
        - {name: recruitmentId, in: path, required: true, schema: {type: integer, format: int32}}
      responses:
        '200': {description: 申请已提交, content: {application/json: {schema: {$ref: '#/components/schemas/IntegerResult'}}}}
        default: {$ref: '#/components/responses/ErrorResponse'}
  /team/{teamId}/applications:
    get:
      tags: [人才招募]
      operationId: listTeamApplications
      summary: 查询队伍收到的入队申请
      security: [{bearerAuth: []}]
      x-required-roles: [TEAM_LEADER, EVENT_ADMIN]
      parameters: [{$ref: '#/components/parameters/TeamId'}]
      responses:
        '200': {description: 获取成功, content: {application/json: {schema: {$ref: '#/components/schemas/TeamApplicationListResult'}}}}
        default: {$ref: '#/components/responses/ErrorResponse'}
  /team/{teamId}/applications/{applicationId}/accept:
    post:
      tags: [人才招募]
      operationId: acceptTeamApplication
      summary: 接受入队申请
      security: [{bearerAuth: []}]
      x-required-roles: [TEAM_LEADER, EVENT_ADMIN]
      parameters:
        - {$ref: '#/components/parameters/TeamId'}
        - {name: applicationId, in: path, required: true, schema: {type: integer, format: int32}}
      responses:
        '200': {$ref: '#/components/responses/OkResponse'}
        default: {$ref: '#/components/responses/ErrorResponse'}
  /team/{teamId}/applications/{applicationId}/reject:
    post:
      tags: [人才招募]
      operationId: rejectTeamApplication
      summary: 拒绝入队申请
      security: [{bearerAuth: []}]
      x-required-roles: [TEAM_LEADER, EVENT_ADMIN]
      parameters:
        - {$ref: '#/components/parameters/TeamId'}
        - {name: applicationId, in: path, required: true, schema: {type: integer, format: int32}}
      responses:
        '200': {$ref: '#/components/responses/OkResponse'}
        default: {$ref: '#/components/responses/ErrorResponse'}
  /team/events/{eventId}/recommendation:
    parameters: [{$ref: '#/components/parameters/EventId'}]
    put:
      tags: [人才招募]
      operationId: publishMyRecommendation
      summary: 发布或重新开放本人自荐名片
      security: [{bearerAuth: []}]
      x-required-roles: [STUDENT]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/PublishRecommendationDTO'}}}
      responses:
        '200': {description: 自荐已发布, content: {application/json: {schema: {$ref: '#/components/schemas/IntegerResult'}}}}
        default: {$ref: '#/components/responses/ErrorResponse'}
    delete:
      tags: [人才招募]
      operationId: closeMyRecommendation
      summary: 撤下本人自荐名片
      security: [{bearerAuth: []}]
      x-required-roles: [STUDENT]
      responses:
        '200': {$ref: '#/components/responses/OkResponse'}
        default: {$ref: '#/components/responses/ErrorResponse'}
  /team/events/{eventId}/recommendations/list:
    post:
      tags: [人才招募]
      operationId: listStudentRecommendations
      summary: 分页检索未组队选手自荐名片
      security: [{bearerAuth: []}]
      x-required-roles: [STUDENT]
      parameters:
        - {$ref: '#/components/parameters/EventId'}
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/Size'}
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/QueryRecommendationDTO'}}}
      responses:
        '200': {description: 获取成功, content: {application/json: {schema: {$ref: '#/components/schemas/RecommendationPageResult'}}}}
        default: {$ref: '#/components/responses/ErrorResponse'}
  /team/events/{eventId}/admin/list:
    post:
      tags: [队伍]
      operationId: adminListTeams
      summary: 赛事管理员分页检索队伍
      security: [{bearerAuth: []}]
      x-required-roles: [EVENT_ADMIN]
      parameters:
        - {$ref: '#/components/parameters/EventId'}
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/Size'}
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/QueryTeamDTO'}}}
      responses:
        '200': {description: 获取成功, content: {application/json: {schema: {$ref: '#/components/schemas/TeamPageResult'}}}}
        default: {$ref: '#/components/responses/ErrorResponse'}
  /submission/phase/{phaseId}/team/{teamId}:
    parameters:
      - {$ref: '#/components/parameters/PhaseId'}
      - {$ref: '#/components/parameters/TeamId'}
    get:
      tags: [作品]
      operationId: getSubmissionByPhaseAndTeam
      summary: 获取队伍在轮次中的作品
      description: 尚未提交时，data 仍包含本轮动态表单配置，其余作品字段可为空。
      security: [{bearerAuth: []}]
      x-required-roles: [TEAM_MEMBER, EVENT_ADMIN, EVENT_JUDGE]
      responses:
        '200': {description: 获取成功, content: {application/json: {schema: {$ref: '#/components/schemas/SubmissionInfoResult'}}}}
        default: {$ref: '#/components/responses/ErrorResponse'}
    post:
      tags: [作品]
      operationId: submitWork
      summary: 提交或重新提交作品
      description: 每次提交生成一个新版本；实际必填项由轮次提交配置动态决定。
      security: [{bearerAuth: []}]
      x-required-roles: [TEAM_LEADER]
      x-required-window: SUBMIT
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/SubmitWorkDTO'}}}
      responses:
        '200': {description: 提交成功, content: {application/json: {schema: {$ref: '#/components/schemas/SubmissionIdResult'}}}}
        default: {$ref: '#/components/responses/ErrorResponse'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 登录或兑换后返回的 ACCESS 类型 JWT。
  parameters:
    EventId: {name: eventId, in: path, required: true, schema: {type: integer, format: int32, minimum: 1}, description: 赛事 ID}
...
    PhaseId: {name: phaseId, in: path, required: true, schema: {type: integer, format: int32, minimum: 1}, description: 轮次 ID}
    TeamId: {name: teamId, in: path, required: true, schema: {type: integer, format: int32, minimum: 1}, description: 队伍 ID}
...
    Page: {name: page, in: query, required: false, schema: {type: integer, minimum: 1, default: 1}, description: 页码，从 1 开始}
    Size: {name: size, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 100, default: 10}, description: 每页条数}
  responses:
    OkResponse:
      description: 操作成功
      content:
        application/json:
          schema: {$ref: '#/components/schemas/VoidResult'}
    ErrorResponse:
      description: 业务异常、鉴权失败、参数错误或系统异常
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResult'}
          examples:
            parameterError: {summary: 参数错误, value: {code: 4000, data: null, msg: 参数错误}}
            unauthorized: {summary: 未携带令牌, value: {code: 1004, data: null, msg: 未携带token}}
            notFound: {summary: 资源不存在, value: {code: 3001, data: null, msg: 赛事不存在}}
  schemas:
    LocalDateTime:
      type: string
      description: ISO-8601 本地日期时间，不包含时区偏移；服务端业务时区为 Asia/Shanghai。
      pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(\.\d+)?$'
      example: '2026-08-30T14:30:00'
...
    TeamType: {type: string, enum: [同专业, 跨专业, 跨校区]}
    TeamStatus: {type: string, enum: [晋级, 淘汰]}
...
    TeamFlowStatus: {type: string, enum: [待处理, 已接受, 已拒绝, 已取消]}
...
    UserIdDTO:
      type: object
      required: [userId]
      properties:
        userId: {type: integer, format: int32}
...
    CreateTeamDTO:
      type: object
      required: [name, type]
      properties:
        name: {type: string, minLength: 1}
        introduction: {type: string}
        type: {$ref: '#/components/schemas/TeamType'}
    UpdateTeamDTO:
      type: object
      required: [name, type, version]
      properties:
        name: {type: string, minLength: 1}
        introduction: {type: string}
        type: {$ref: '#/components/schemas/TeamType'}
        version: {type: integer, format: int32}
...
    InviteCodeDTO:
      type: object
      required: [code]
      properties:
        code: {type: string, pattern: '^[A-Za-z0-9]{12}$'}
    CreateRecruitmentDTO:
      type: object
      required: [title, vacancies]
      properties:
        title: {type: string, minLength: 1, maxLength: 100}
        description: {type: string, maxLength: 1000}
        requiredTags: {type: array, maxItems: 20, items: {type: string}}
        vacancies: {type: integer, minimum: 1}
    UpdateRecruitmentDTO:
      allOf:
        - {$ref: '#/components/schemas/CreateRecruitmentDTO'}
        - type: object
          required: [open, version]
          properties:
            open: {type: boolean}
            version: {type: integer, format: int32}
    QueryRecruitmentDTO:
      type: object
      properties:
        trackId: {type: integer, format: int32}
        keyword: {type: string}
        skill: {type: string}
    PublishRecommendationDTO:
      type: object
      required: [introduction]
      properties:
        introduction: {type: string, minLength: 1, maxLength: 1000}
        skills: {type: array, maxItems: 20, items: {type: string}}
    QueryRecommendationDTO:
      type: object
      properties:
        trackId: {type: integer, format: int32}
        keyword: {type: string}
        skill: {type: string}
    QueryTeamDTO:
      type: object
      properties:
        trackId: {type: integer, format: int32}
        name: {type: string}
        status: {$ref: '#/components/schemas/TeamStatus'}
    SubmitWorkDTO:
      type: object
      description: 字段是否必填由轮次 submissionConfig 动态决定。
      properties:
        repoUrl: {type: string, maxLength: 255}
        licenseType: {type: string, maxLength: 50}
        derivedFrom: {type: string, maxLength: 500}
        archiveFileId: {type: integer, format: int64, minimum: 1}
        videoFileId: {type: integer, format: int64, minimum: 1}
        videoUrl: {type: string, maxLength: 500}
        docFileId: {type: integer, format: int64, minimum: 1}
        demoUrl: {type: string, maxLength: 500}
        introMd: {type: string}
        declaration: {type: string}
        changeLog: {type: string, maxLength: 200}
        version: {type: integer, format: int32, description: 乐观锁版本号；首次提交可省略}
...
    TeamIdVO:
      type: object
      properties:
        teamId: {type: integer, format: int32}
        inviteCode: {type: string, description: 创建时生成的12位私密邀请码}
...
    TeamMemberVO:
      type: object
      properties:
        userId: {type: integer, format: int32}
        name: {type: string}
        casId: {type: [string, 'null']}
        campus: {type: [string, 'null']}
        major: {type: [string, 'null']}
        tags: {type: array, items: {type: string}}
        leader: {type: boolean}
    TeamInfoVO:
      type: object
      properties:
        teamId: {type: integer, format: int32}
        name: {type: string}
        eventId: {type: integer, format: int32}
        eventName: {type: string}
        trackId: {type: integer, format: int32}
        trackName: {type: string}
        size: {type: integer, format: int32}
        introduction: {type: [string, 'null']}
        type: {$ref: '#/components/schemas/TeamType'}
        inviteCode: {type: string, description: 仅队员和赛事管理员可见}
        version: {type: integer, format: int32}
        members: {type: array, items: {$ref: '#/components/schemas/TeamMemberVO'}}
...
    TeamApplicationVO:
      type: object
      properties:
        applicationId: {type: integer, format: int32}
        recruitmentId: {type: integer, format: int32}
        teamId: {type: integer, format: int32}
        userId: {type: integer, format: int32}
        userName: {type: string}
        campus: {type: string}
        major: {type: string}
        tags: {type: array, items: {type: string}}
        status: {$ref: '#/components/schemas/TeamFlowStatus'}
        updateTime: {$ref: '#/components/schemas/LocalDateTime'}
    TeamInvitationVO:
      type: object
      properties:
        invitationId: {type: integer, format: int32}
        teamId: {type: integer, format: int32}
        teamName: {type: string}
        eventId: {type: integer, format: int32}
        eventName: {type: string}
        trackId: {type: integer, format: int32}
        trackName: {type: string}
        inviterId: {type: integer, format: int32}
        inviterName: {type: string}
        status: {$ref: '#/components/schemas/TeamFlowStatus'}
        updateTime: {$ref: '#/components/schemas/LocalDateTime'}
...
    TeamBriefVO:
      type: object
      properties:
        teamId: {type: integer, format: int32}
        name: {type: string}
        trackId: {type: integer, format: int32}
        trackName: {type: string}
        leaderId: {type: integer, format: int32}
        leaderName: {type: string}
        size: {type: integer}
        type: {$ref: '#/components/schemas/TeamType'}
        status: {$ref: '#/components/schemas/TeamStatus'}
        updateTime: {$ref: '#/components/schemas/LocalDateTime'}
...
    SubmissionIdVO:
      type: object
      properties:
        submissionId: {type: integer, format: int32}
        versionNo: {type: integer, format: int32}
...
    RecruitmentPage:
      type: object
      properties:
        records: {type: array, items: {$ref: '#/components/schemas/RecruitmentVO'}}
        total: {type: integer, format: int64}
        size: {type: integer, format: int64}
        current: {type: integer, format: int64}
        pages: {type: integer, format: int64}
    RecommendationPage:
      type: object
      properties:
        records: {type: array, items: {$ref: '#/components/schemas/RecommendationVO'}}
        total: {type: integer, format: int64}
        size: {type: integer, format: int64}
        current: {type: integer, format: int64}
        pages: {type: integer, format: int64}
    TeamPage:
      type: object
      properties:
        records: {type: array, items: {$ref: '#/components/schemas/TeamBriefVO'}}
        total: {type: integer, format: int64}
        size: {type: integer, format: int64}
        current: {type: integer, format: int64}
        pages: {type: integer, format: int64}
...
    ResultBase:
      type: object
      required: [code, data, msg]
      properties:
        code: {type: integer, format: int32, description: 成功为 200，失败为业务状态码}
        data: {}
        msg: {type: string}
    VoidResult:
      allOf: [{$ref: '#/components/schemas/ResultBase'}, {type: object, properties: {data: {type: 'null'}}}]
      example: {code: 200, data: null, msg: 成功}
    ErrorResult:
      allOf: [{$ref: '#/components/schemas/ResultBase'}, {type: object, properties: {data: {type: 'null'}}}]
    StringResult: {allOf: [{$ref: '#/components/schemas/ResultBase'}, {type: object, properties: {data: {type: string}}}]}
    IntegerResult: {allOf: [{$ref: '#/components/schemas/ResultBase'}, {type: object, properties: {data: {type: integer, format: int32}}}]}
...
    TeamIdResult: {allOf: [{$ref: '#/components/schemas/ResultBase'}, {type: object, properties: {data: {$ref: '#/components/schemas/TeamIdVO'}}}]}
    TeamInfoResult: {allOf: [{$ref: '#/components/schemas/ResultBase'}, {type: object, properties: {data: {$ref: '#/components/schemas/TeamInfoVO'}}}]}
    TeamInvitationListResult: {allOf: [{$ref: '#/components/schemas/ResultBase'}, {type: object, properties: {data: {type: array, items: {$ref: '#/components/schemas/TeamInvitationVO'}}}}]}
    TeamApplicationListResult: {allOf: [{$ref: '#/components/schemas/ResultBase'}, {type: object, properties: {data: {type: array, items: {$ref: '#/components/schemas/TeamApplicationVO'}}}}]}
    RecruitmentPageResult: {allOf: [{$ref: '#/components/schemas/ResultBase'}, {type: object, properties: {data: {$ref: '#/components/schemas/RecruitmentPage'}}}]}
    RecommendationPageResult: {allOf: [{$ref: '#/components/schemas/ResultBase'}, {type: object, properties: {data: {$ref: '#/components/schemas/RecommendationPage'}}}]}
    TeamPageResult: {allOf: [{$ref: '#/components/schemas/ResultBase'}, {type: object, properties: {data: {$ref: '#/components/schemas/TeamPage'}}}]}
    SubmissionInfoResult: {allOf: [{$ref: '#/components/schemas/ResultBase'}, {type: object, properties: {data: {$ref: '#/components/schemas/SubmissionInfoVO'}}}]}
    SubmissionIdResult: {allOf: [{$ref: '#/components/schemas/ResultBase'}, {type: object, properties: {data: {$ref: '#/components/schemas/SubmissionIdVO'}}}]}
...